Refactor model & response data
