+++ lib/main.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:get/get.dart';
-import 'package:survey/api/api_client_provider.dart';
+import 'package:survey/api/http/api_client_provider.dart';
 import 'package:survey/navigator/navigator.dart';
 import 'package:survey/pages/home/home_page.dart';
 import 'package:survey/pages/login/login_binding.dart';
@@ -12,7 +12,7 @@ import 'package:survey/pages/splash/splash_page.dart';
 import 'package:survey/preferences/shared_preferences.dart';
 import 'package:survey/themes.dart';
 
-import 'api/api_client.dart';
+import 'api/http/api_client.dart';
 import 'flavors.dart';
 
 void main() {
