+++ lib/use_cases/get_login_state_use_case.dart
@@ -8,8 +8,9 @@ class GetLoginStateUseCase extends NoParamsUseCase<bool> {
 
   @override
   Future<Result<bool>> call() {
-    return _sharedPreferencesStorage
-        .getAccessToken()
-        .then((token) => Success(token != null && token.isNotEmpty));
+    return _sharedPreferencesStorage.getAccessToken().then((token) {
+      print(token);
+      return Success(token != null && token.isNotEmpty);
+    });
   }
 }
