+++ pubspec.yaml
@@ -28,7 +28,7 @@ dependencies:
     sdk: flutter
   intl: ^0.17.0
 
-  shared_preferences: ^2.0.5
+  shared_preferences: ^0.5.12+4
 
   flutter_svg: ^0.19.3
 
@@ -46,6 +46,8 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   optional: ^5.0.0+1
+
+  graphql_flutter: ^4.0.1
 
 dev_dependencies:
   flutter_test:
